blog/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
gunicorn==20.1.0
Markdown==3.3.4
nltk==3.6.5
pytz==2021.3
regex==2021.11.10
